chore: make .index() and .scope() use accessors (#480)

+++ crates/grit-pattern-matcher/src/constants.rs
@@ -3,7 +3,7 @@ pub const PROGRAM_INDEX: usize = 1;
 pub const FILENAME_INDEX: usize = 2;
 pub const ABSOLUTE_PATH_INDEX: usize = 3;
 
-pub const GLOBAL_VARS_SCOPE_INDEX: usize = 0;
+pub const GLOBAL_VARS_SCOPE_INDEX: u16 = 0;
 
 pub const MATCH_VAR: &str = "$match";
 pub const GRIT_RANGE_VAR: &str = "$grit_range";
